+++ iOSPhotosScreenHW14/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
