+++ iOSPhotosScreenHW14/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
