+++ iOSPhotosScreenHW14/Resources/Assets.xcassets/Contents.json
