+++ iOSPhotosScreenHW14/Sources/AlbumsLayout/AlbumsLayoutViewController.swift
@@ -7,13 +7,9 @@
 
 import UIKit
 
-class ViewController: UIViewController {
+class AlbumsLayoutViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        // Do any additional setup after loading the view.
     }
-
-
 }
-
