+++ iOSPhotosScreenHW14/Sources/App/AppDelegate.swift
@@ -1,0 +1,16 @@
+//
+//  AppDelegate.swift
+//  iOSPhotosScreenHW14
+//
+//  Created by Артур Горлов on 06.09.2022.
+//
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        return true
+    }
+}
