+++ iOSPhotosScreenHW14/Sources/App/SceneDelegate.swift
@@ -1,0 +1,17 @@
+//
+//  SceneDelegate.swift
+//  iOSPhotosScreenHW14
+//
+//  Created by Артур Горлов on 06.09.2022.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let _ = (scene as? UIWindowScene) else { return }
+    }
+}
